requirements.txt
pyautogen
openai
python-dotenv
tiktoken
pydantic
langchain
chromadb
unstructured
pdfminer.six
python-docx
pypdf
nltk
jsonschema
rich
loguru
pandas
evaluate
rouge-score
